Refactor model mapping
