+++ tests/panoramic/cli/mapper/test_model.py
